requirements.txt
Django~=4.2
asgiref~=3.6.0
channels~=3.0.4
djangorestframework~=3.14.0